Create new StairwayLoggingHooks class that implements the StairwayHook Interface
